third_party/WebKit/LayoutTests/accessibility/table-with-presentation-role.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id="container">
  <table role="presentation">
    <tr>
      <td>A</td><td>B</td>
    </tr>
    <tr>
      <td>C</td><td>D</td>
    </tr>
  </table>
</div>

<script>
test(() => {
  var axContainer = accessibilityController.accessibleElementById('container');
  assert_equals(axContainer.childrenCount, 4);
  var axText1 = axContainer.childAtIndex(0);
  assert_equals(axText1.role, "AXRole: AXStaticText");
  assert_equals(axText1.name, "A");
  var axText2 = axContainer.childAtIndex(1);
  assert_equals(axText2.role, "AXRole: AXStaticText");
  assert_equals(axText2.name, "B");
}, "ARIA presentation role on table element has no rows or cells.");

if (window.testRunner)
    document.getElementById("container").style.display = "none";
</script>
